Align package names in :instrumentation-agent-services
