+++ platforms/core-runtime/instrumentation-agent-services/src/integTest/groovy/org/gradle/internal/instrumentation/agent/AgentApplicationTest.groovy
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package org.gradle.instrumentation.agent
+package org.gradle.internal.instrumentation.agent
 
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.configurationcache.ConfigurationCacheFixture
 import org.gradle.integtests.fixtures.daemon.DaemonLogsAnalyzer
 import org.gradle.integtests.fixtures.executer.GradleHandle
-import org.gradle.internal.agents.AgentStatus
 import org.gradle.launcher.daemon.configuration.DaemonBuildOptions
 import org.gradle.test.precondition.Requires
 import org.gradle.test.preconditions.IntegTestPreconditions
