+++ platforms/core-runtime/instrumentation-agent-services/src/main/java/org/gradle/internal/instrumentation/agent/DisabledAgentStatus.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 the original author or authors.
+ * Copyright 2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,5 +14,11 @@
  * limitations under the License.
  */
 
-@org.gradle.api.NonNullApi
-package org.gradle.internal.agents;
+package org.gradle.internal.instrumentation.agent;
+
+class DisabledAgentStatus implements AgentStatus {
+    @Override
+    public boolean isAgentInstrumentationEnabled() {
+        return false;
+    }
+}
