+++ platforms/extensibility/unit-test-fixtures/src/main/java/org/gradle/testfixtures/internal/TestGlobalScopeServices.java
@@ -18,8 +18,8 @@ package org.gradle.testfixtures.internal;
 import org.gradle.cache.FileLockManager;
 import org.gradle.cache.internal.CacheFactory;
 import org.gradle.internal.Factory;
-import org.gradle.internal.agents.AgentStatus;
 import org.gradle.internal.concurrent.ExecutorFactory;
+import org.gradle.internal.instrumentation.agent.AgentStatus;
 import org.gradle.internal.logging.LoggingManagerInternal;
 import org.gradle.internal.operations.BuildOperationListenerManager;
 import org.gradle.internal.operations.BuildOperationProgressEventEmitter;
